refactor(explorer): migrate TRANSITIONS to PanelBuilder registry
Phase 2/8 (6th of 9 PlotTypes total).

* New explorer_internals/transitions.py hosts
  TransitionsPanelBuilder + the absorbed
  display_transitions render function (largest builder so far,
  ~100 lines of widget construction, ~40 lines of dispatch).
* The self.ui["Transitions"] nested-dict state is bound via
  a local transitions_ui alias to make the per-widget
  assignments more scannable.
* Registry adds PlotType.TRANSITIONS; the matching dispatch
  branches are removed from Explorer.build_panel and
  ExplorerSettings.build_settings_ui.
* The bare_dressed_toggle observer callback remains on the
  Explorer class itself — it mutates the
  initial_dressed_inttext widget's display style, which is
  cross-builder UI concern that does not belong inside the
  TRANSITIONS builder.
