+++ scqubits/explorer/explorer_settings.py
@@ -111,112 +111,6 @@ class ExplorerSettings:
         subsys: Any = plot_id.subsystems
         plot_type = plot_id.plot_type
 
-        if plot_type is PlotType.TRANSITIONS:
-            self.ui["Transitions"]["initial_state_inttexts"] = [
-                ui.ValidatedNumberField(
-                    label=subsys.id_str,
-                    num_type=int,
-                    v_min=0,
-                    v_max=subsys.truncated_dim,
-                    v_model=0,
-                    style_="display: inherit; width: 65px;",
-                    class_="ml-4",
-                )
-                for subsys in self.explorer.sweep.hilbertspace
-            ]
-
-            self.ui["Transitions"]["initial_dressed_inttext"] = ui.ValidatedNumberField(
-                label="Dressed state",
-                class_="ml-4 align-bottom",
-                num_type=int,
-                v_min=0,
-                v_max=self.explorer.sweep.hilbertspace.dimension,
-                v_model=0,
-                style_="display: none; width: 65px;",
-            )
-
-            self.ui["Transitions"]["photons_inttext"] = ui.ValidatedNumberField(
-                num_type=int,
-                class_="ml-3",
-                v_model=1,
-                v_min=1,
-                v_max=5,
-                label="Photon number",
-                style_="max-width: 120px",
-            )
-            self.ui["Transitions"]["highlight_selectmultiple"] = ui.InitializedSelect(
-                multiple=True,
-                label="",
-                items=self.explorer.subsys_names,
-                v_model=[self.explorer.subsys_names[0]],
-                width=185,
-            )
-
-            self.ui["Transitions"]["initial_bare_dressed_toggle"] = v.RadioGroup(
-                v_model="bare",
-                children=[
-                    v.Radio(label="by bare product label", value="bare"),
-                    v.Radio(label="by dressed index", value="dressed"),
-                ],
-            )
-
-            self.ui["Transitions"]["sidebands_switch"] = v.Switch(
-                label="Show sidebands", v_model=False, width=250
-            )
-            for inttext in self.ui["Transitions"]["initial_state_inttexts"]:
-                inttext.observe(self.explorer.update_plots, names="num_value")
-            self.ui["Transitions"]["initial_dressed_inttext"].observe(
-                self.explorer.update_plots, names="num_value"
-            )
-            self.ui["Transitions"]["photons_inttext"].observe(
-                self.explorer.update_plots, names="num_value"
-            )
-            self.ui["Transitions"]["highlight_selectmultiple"].observe(
-                self.explorer.update_plots, names="v_model"
-            )
-            self.ui["Transitions"]["sidebands_switch"].observe(
-                self.explorer.update_plots, names="v_model"
-            )
-            self.ui["Transitions"]["initial_bare_dressed_toggle"].observe(
-                self.explorer.bare_dressed_toggle, names="v_model"
-            )
-
-            initial_state_selection = ui.flex_row(
-                [
-                    v.Text(children=["Initial state"]),
-                    self.ui["Transitions"]["initial_bare_dressed_toggle"],
-                    *self.ui["Transitions"]["initial_state_inttexts"],
-                    self.ui["Transitions"]["initial_dressed_inttext"],
-                ]
-            )
-
-            photon_options_selection = v.Container(
-                class_="d-flex flex-row",
-                children=[
-                    v.Text(children=["Single vs. multi-photon transitions"]),
-                    self.ui["Transitions"]["photons_inttext"],
-                ],
-            )
-            transition_highlighting = v.Container(
-                class_="d-flex flex-row",
-                children=[
-                    v.Text(children=["Highlight:"]),
-                    self.ui["Transitions"]["highlight_selectmultiple"],
-                ],
-            )
-
-            return [
-                v.Container(
-                    class_="d-flex flex-column",
-                    children=[
-                        initial_state_selection,
-                        photon_options_selection,
-                        self.ui["Transitions"]["sidebands_switch"],
-                        transition_highlighting,
-                    ],
-                ),
-            ]
-
         if plot_type is PlotType.SELF_KERR and isinstance(subsys[0], scq.Oscillator):
             return []
